pyproject.toml
[project]
name = "cad-diagram-analyzer"
version = "0.1.0"
description = "Agentic CAD diagram analysis pipeline using multimodal LLMs and multi-resolution tiling"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # Core data / config
    "pydantic>=2.10",
    "python-dotenv>=1.0",

    # GCP services
    "google-cloud-storage>=2.18",
    "google-cloud-firestore>=2.19",
    "google-cloud-documentai>=2.29",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",

    # Agent framework + Gemini SDK
    "google-adk>=1.0.0",
    "google-genai>=1.9.0",

    # Image processing
    "Pillow>=10.4",
    "opencv-python-headless>=4.10",
    "numpy>=1.26",

    # Web server (Phase 5)
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",

    # Structured logging
    "structlog>=24.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.23",
    "mypy>=1.15",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ---------------------------------------------------------------------------
# Ruff — linting + formatting
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
# E501 line-too-long: handled by formatter, not lint
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
# __init__.py re-exports are fine to have without explicit use
"__init__.py" = ["F401"]

# ---------------------------------------------------------------------------
# Mypy — static type checking
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
explicit_package_bases = true
# Pydantic v2 plugin keeps mypy happy with model fields
plugins = ["pydantic.mypy"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"
